refactor: fractional digits
